requirements.txt
python-dotenv>=0.15.0
psycopg2>=2.8.6
requests>=2.25.1